Add support for .NET Standard 1.6
This work migrates the C# bindings to project.json based projects in
order to build assemblies for both .NET Standard 1.6 and .NET Framework
4.6.2.

Notes:

1. It was necessary to use .NET Framework 4.6.2 because a method
available in .NET Core was not available in .NET Framework 4.5.
2. Certain features are not supported for .NET Core as the required
functionality is lacking from the framework:
- Datatables: System.Data.DataTable not fully supported in .NET Standard
1.6
- XML validation:  System.Xml.XmlSchema not fully supported in .NET
Standard 1.6
- Serialization of VariantException.

One other item of note is that it was necessary to make XmlWriter
disposable. This was because it was no longer possible to use
XmlTextWriter and the replacement required disposal: without disposal
nothing was written. As most usage appears to be via the static method
XmlWriter.ToString(...) which has been changed to dispose the writer it
creates, this should not be too much of an issue.

The version has been bumped to 2.0.0. Aside from anything that means we
can now fix the assembly file version at 2.0.0.0 to avoid dependency
issues in subsequent releases.